+++ meetup2apricot/meetup_event.py
@@ -30,7 +30,7 @@ YES_RSVP_COUNT = "yes_rsvp_count"
 THREE_HOURS_MSEC = 3 * 60 * 60 * 1000
 DEFAULT_DURATION = THREE_HOURS_MSEC
 
-ACCOUNTING_CODE_PATTERN = re.compile(r"([A-Z3][A-Z]*)(?:_[A-Z_]*)?:")
+ACCOUNTING_CODE_PATTERN = re.compile(r"(?:([A-Z3][A-Z]*)[_:]+)")
 MEMBER_ONLY_PATTERN = re.compile(r"members[ -]*only", re.IGNORECASE)
 
 MeetupVenue = namedtuple("MeetupVenue", "name address city state zipcode")
@@ -128,13 +128,9 @@ class MeetupEvent:
         return self.event_json.get(YES_RSVP_COUNT, 0)
 
     @property
-    def accounting_code(self):
-        """Return the accounting code that prefixes the event name."""
-        match = ACCOUNTING_CODE_PATTERN.match(self.name)
-        if match:
-            return match.group(1)
-        else:
-            return None
+    def accounting_codes(self):
+        """Return a list of accounting codes that prefix the event name."""
+        return ACCOUNTING_CODE_PATTERN.findall(self.name)
 
     @property
     def status(self):